remove live api + filter empty apprenticeships
