+++ layouts/partials/home/contributors.html
@@ -8,7 +8,16 @@
         <p class="mt-12 mb-12">
           This site and the listings are all available because of contributors like you. If you’re looking to contribute to any of the listings or to this site, take a look at the contributing guidelines in the repository below.
         </p>
-        <ul id="contributor-list" class="list-reset flex flex-wrap justify-center"></ul>
+        <ul id="contributor-list" class="list-reset flex flex-wrap justify-center">
+
+          {{range .Site.Data.github }}
+            <li class= "m-2 rounded-full shadow hover-shadow-lg transition">
+              <a href="{{.url}}" target="_blank" rel="nooppener">
+                <img class="w-24 h-24 rounded-full" src="{{ .avatar }}" alt="{{.user}}" />
+              </a>
+            </li>
+          {{ end }}
+        </ul>
       </div>
     </div>
 </section>
