+++ static/index.js
@@ -1,4 +1,1 @@
-// var instantsearch = require("instantsearch.js");
 require("./components/search");
-require("./components/github");
-// require('./components/scroll-nav');
